Tubonge/requirements.txt
aiosmtplib==3.0.2
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
defusedxml==0.8.0rc2
Deprecated==1.2.13
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.5.0
django-crontab==0.7.1
django-filter==24.3
django-rest-auth==0.9.5
django-smtp-ssl==1.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-swagger==0.2.2
djoser==2.2.3
drf-spectacular==0.27.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.26.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.12.48
pkce==1.0.3
pycparser==2.22
pycryptodome==3.10.4
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tldextract==3.1.0
twilio==7.9.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
whitelist==0.2.8
whitenoise==6.7.0
wrapt==1.16.0
